+++ MercadoPagoSDK/MercadoPagoSDK/Images.xcassets/CardSlider/Overlay_.imageset/Contents.json
