+++ MercadoPagoSDK/MercadoPagoSDK/Images.xcassets/CardSlider/Overlay_.imageset/overlay@2x.png
