+++ MercadoPagoSDK/MercadoPagoSDK/Images.xcassets/CardSlider/Overlay_.imageset/overlay@3x.png
