+++ MercadoPagoSDK/MercadoPagoSDK/UI/PXCard/DefaultCards/AccountMoneyCard.swift
@@ -1,0 +1,61 @@
+//
+//  AccountMoneyCard.swift
+//  MercadoPagoSDK
+//
+//  Created by Juan sebastian Sanzone on 17/10/18.
+//
+
+import Foundation
+
+class AccountMoneyCard: NSObject, CardUI {
+    var placeholderName = ""
+    var placeholderExpiration = ""
+    var bankImage: UIImage?
+    var cardPattern = [0]
+    var cardFontColor: UIColor = UIColor(red: 105/255, green: 105/255, blue: 105/255, alpha: 1)
+    var cardLogoImage: UIImage?
+    var cardBackgroundColor: UIColor = UIColor(red:0.00, green:0.64, blue:0.85, alpha:1.0)
+    var securityCodeLocation: Location = .back
+    var defaultUI = false
+    var securityCodePattern = 3
+    var fontType: String = "light"
+}
+
+extension AccountMoneyCard {
+    static func render(containerView: UIView, balanceText: String) {
+        let amImage = UIImageView()
+        amImage.backgroundColor = .clear
+        amImage.contentMode = .scaleAspectFit
+        amImage.image = ResourceManager.shared.getImage("amImage")
+        amImage.alpha = 0.6
+        containerView.addSubview(amImage)
+        PXLayout.setWidth(owner: amImage, width: PXCardSliderSizeManager.getItemContainerSize().height * 0.65).isActive = true
+        PXLayout.setHeight(owner: amImage, height: PXCardSliderSizeManager.getItemContainerSize().height * 0.65).isActive = true
+        PXLayout.pinTop(view: amImage).isActive =  true
+        PXLayout.pinRight(view: amImage).isActive = true
+
+        let patternView = UIImageView()
+        patternView.contentMode = .scaleAspectFit
+        patternView.image = ResourceManager.shared.getImage("amPattern")
+        let patternDeltaOffset: CGFloat = -40
+        containerView.addSubview(patternView)
+        PXLayout.setWidth(owner: patternView, width: PXCardSliderSizeManager.getItemContainerSize().width).isActive = true
+        PXLayout.setHeight(owner: patternView, height: PXCardSliderSizeManager.getItemContainerSize().height).isActive = true
+        PXLayout.pinTop(view: patternView).isActive = true
+        PXLayout.pinRight(view: patternView, withMargin: patternDeltaOffset).isActive = true
+
+        let label = UILabel()
+        label.text = " \(balanceText)   " //TODO: Proper inset with labels.
+        label.font = Utils.getFont(size: PXLayout.XXS_FONT)
+        label.textColor = UIColor.white
+        label.textAlignment = .center
+        label.layer.cornerRadius = 5
+        label.layer.borderWidth = 1
+        label.layer.borderColor = UIColor(red:1.00, green:1.00, blue:1.00, alpha:0.5).cgColor
+        containerView.addSubview(label)
+
+        PXLayout.pinLeft(view: label, withMargin: PXLayout.S_MARGIN).isActive = true
+        PXLayout.pinBottom(view: label, withMargin: PXLayout.S_MARGIN).isActive = true
+        PXLayout.setHeight(owner: label, height: PXLayout.L_MARGIN).isActive = true
+    }
+}
