+++ MercadoPagoSDK/MercadoPagoSDK/UI/PXCard/View/UIView + Nib.swift
@@ -1,6 +1,6 @@
 import UIKit
 
-extension UIView {
+internal extension UIView {
 
     func loadFromNib() {
